reorganizes files to adhere to best practices and get rid of cruft/extra files
