--- src/styles/ArticleContainer.css
@@ -1,22 +1,0 @@
-@import url(accessories.css);
-
-.no-articles {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin: 12px;
-    padding: 15px;
-    border: 1px solid var(--item);
-    border-radius: 5px;
-    background-color: var(--item);
-}
-
-.no-articles > * {
-    color: var(--text-dark);
-    margin: 15px 0px;
-}
-
-.no-articles > h3 {
-    margin: 12px 0px;
-}
