+++ src/styles/SearchPage.css
@@ -1,0 +1,61 @@
+@import url(accessories.css);
+
+/* === SEARCH STYLING === */
+.search-container > form {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    background-color: var(--item);
+    border: 2px solid var(--item);
+    border-radius: 5px;
+    margin: 12px;
+    padding: 5px;
+    font-size: var(--font-size);
+}
+
+/* Search input */
+.search-container > form > input {
+    padding: 15px;
+    margin-right: 10px;
+    border: none;
+    border-radius: 5px;
+    width: 90%;
+    font-size: 18px;
+    color: var(--text-dark);
+}
+
+.search-container > form > input:hover {
+    transform: scale(1.005);
+    transition: 0.3s;
+}
+
+.search-container > form > input:focus {
+    outline: none;
+}
+
+/* Search Buttons */
+.search-btns {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+}
+
+.search-btns > button {
+    color: var(--text-dark);
+    background-color: var(--item);
+    border: solid 1px var(--item-border);
+    border-radius: 5px;
+    margin: 0px 10px;
+    padding: 10px;
+    cursor: pointer;
+    white-space: nowrap;
+    transition: 0.2s;
+    font-size: var(--font-size-alt);
+}
+
+.search-btns > button:hover {
+    color: var(--text-scarlet);
+    background-color: var(--item-hover);
+}
